RENAME DTO request와 response 폴더로 분할 관리
