+++ src/main/java/com/jeontongju/authentication/dto/request/SellerInfoForSignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.jeontongju.authentication.dto;
+package com.jeontongju.authentication.dto.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotNull;
