+++ src/main/java/com/jeontongju/authentication/dto/response/ImpAuthInfoResponse.java
@@ -1,4 +1,4 @@
-package com.jeontongju.authentication.dto;
+package com.jeontongju.authentication.dto.response;
 
 import lombok.Getter;
 import lombok.Setter;
