+++ src/main/java/com/jeontongju/authentication/service/feign/consumer/ConsumerClientService.java
@@ -1,6 +1,6 @@
 package com.jeontongju.authentication.service.feign.consumer;
 
-import com.jeontongju.authentication.dto.ConsumerInfoForCreateRequestDto;
+import com.jeontongju.authentication.dto.request.ConsumerInfoForCreateRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
