+++ src/main/java/com/jeontongju/authentication/service/feign/seller/SellerServiceClient.java
@@ -1,6 +1,6 @@
 package com.jeontongju.authentication.service.feign.seller;
 
-import com.jeontongju.authentication.dto.SellerInfoForCreateRequestDto;
+import com.jeontongju.authentication.dto.request.SellerInfoForCreateRequestDto;
 import com.jeontongju.authentication.dto.SuccessFeignFormat;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
